backend/requirements.txt
flask
flask_sqlalchemy
flask_cors
psycopg2-binary
flask-jwt-extended
flask-socketio 
openai
python-dotenv